fix : stt complete, api 연결

See merge request s12-final/S12P31S209!112

+++ Android/app/src/main/java/com/example/myapplication/data/dto/response/ClientResponse.kt
@@ -3,10 +3,10 @@ package com.example.myapplication.data.dto.response
 data class ClientResponse(
     val clientId: Int,
     val name: String,
-    val birth : String,
+    val birth: String, // BE 상에서는 LocalDate로 되어있는데 안되면 확인해보기
     val gender: Int,
-    val height: Int,
-    val weight: Int,
+    val height: Float,
+    val weight: Float,
     val image: String?,
     val address: String
 ) {
